added some secuirty

+++ backend/controllers/admin.js
@@ -3,6 +3,12 @@ const User = require("../models/User.model");
 
 exports.getAllUsers = async (req, res) => {
   try {
+    const token = req.body.token;
+    const userId = jwt.verify(token, process.env.JWT_SECRET);
+    const user = await User.findById(userId.id);
+    if (user.role !== "masterAdmin") {
+      return res.status(400).json({ message: "You are not authorized" });
+    }
     const users = await User.find({
       eventId: null,
       role: { $ne: "masterAdmin" },
@@ -15,8 +21,14 @@ exports.getAllUsers = async (req, res) => {
 
 exports.getUserById = async (req, res) => {
   try {
-    const user = await User.findById(req.params.id);
-    res.status(200).json({ user });
+    const token = req.body.token;
+    const userId = jwt.verify(token, process.env.JWT_SECRET);
+    const user = await User.findById(userId.id);
+    if (user.role !== "masterAdmin") {
+      return res.status(400).json({ message: "You are not authorized" });
+    }
+    const usr = await User.findById(req.params.id);
+    res.status(200).json({ usr });
   } catch (error) {
     res.status(500).json({ error: error.message });
   }
@@ -24,10 +36,16 @@ exports.getUserById = async (req, res) => {
 
 exports.makeJuniorAdmin = async (req, res) => {
   try {
-    const user = await User.findByIdAndUpdate(req.params.id, {
+    const token = req.body.token;
+    const userId = jwt.verify(token, process.env.JWT_SECRET);
+    const user = await User.findById(userId.id);
+    if (user.role !== "masterAdmin") {
+      return res.status(400).json({ message: "You are not authorized" });
+    }
+    const usr = await User.findByIdAndUpdate(req.params.id, {
       role: "juniorAdmin",
     });
-    res.status(200).json({ user });
+    res.status(200).json({ usr });
   } catch (error) {
     res.status(500).json({ error: error.message });
   }
@@ -35,10 +53,16 @@ exports.makeJuniorAdmin = async (req, res) => {
 
 exports.makeExecutiveAdmin = async (req, res) => {
   try {
-    const user = await User.findByIdAndUpdate(req.params.id, {
+    const token = req.body.token;
+    const userId = jwt.verify(token, process.env.JWT_SECRET);
+    const user = await User.findById(userId.id);
+    if (user.role !== "masterAdmin") {
+      return res.status(400).json({ message: "You are not authorized" });
+    }
+    const usr = await User.findByIdAndUpdate(req.params.id, {
       role: "executiveAdmin",
     });
-    res.status(200).json({ user });
+    res.status(200).json({ usr });
   } catch (error) {
     res.status(500).json({ error: error.message });
   }
@@ -46,10 +70,16 @@ exports.makeExecutiveAdmin = async (req, res) => {
 
 exports.removeAdmin = async (req, res) => {
   try {
-    const user = await User.findByIdAndUpdate(req.params.id, {
+    const token = req.body.token;
+    const userId = jwt.verify(token, process.env.JWT_SECRET);
+    const user = await User.findById(userId.id);
+    if (user.role !== "masterAdmin") {
+      return res.status(400).json({ message: "You are not authorized" });
+    }
+    const usr = await User.findByIdAndUpdate(req.params.id, {
       role: "user",
     });
-    res.status(200).json({ user });
+    res.status(200).json({ usr });
   } catch (error) {
     res.status(500).json({ error: error.message });
   }
@@ -57,8 +87,14 @@ exports.removeAdmin = async (req, res) => {
 
 exports.editProfile = async (req, res) => {
   try {
-    const user = await User.findByIdAndUpdate(req.params.id, req.body);
-    res.status(200).json({ user });
+    const token = req.body.token;
+    const userId = jwt.verify(token, process.env.JWT_SECRET);
+    const user = await User.findById(userId.id);
+    if (user.role !== "masterAdmin") {
+      return res.status(400).json({ message: "You are not authorized" });
+    }
+    const usr = await User.findByIdAndUpdate(req.params.id, req.body);
+    res.status(200).json({ usr });
   } catch (error) {
     res.status(500).json({ error: error.message });
   }
@@ -66,8 +102,14 @@ exports.editProfile = async (req, res) => {
 
 exports.deleteUser = async (req, res) => {
   try {
-    const user = await User.findByIdAndDelete(req.params.id);
-    res.status(200).json({ user });
+    const token = req.body.token;
+    const userId = jwt.verify(token, process.env.JWT_SECRET);
+    const user = await User.findById(userId.id);
+    if (user.role !== "masterAdmin") {
+      return res.status(400).json({ message: "You are not authorized" });
+    }
+    const usr = await User.findByIdAndDelete(req.params.id);
+    res.status(200).json({ usr });
   } catch (error) {
     res.status(500).json({ error: error.message });
   }
@@ -75,6 +117,12 @@ exports.deleteUser = async (req, res) => {
 
 exports.getAllExecutives = async (req, res) => {
   try {
+    const token = req.body.token;
+    const userId = jwt.verify(token, process.env.JWT_SECRET);
+    const user = await User.findById(userId.id);
+    if (user.role !== "masterAdmin") {
+      return res.status(400).json({ message: "You are not authorized" });
+    }
     const executives = await User.find({ role: "executiveAdmin" });
     res.status(200).json({ executives });
   } catch (error) {
@@ -84,6 +132,12 @@ exports.getAllExecutives = async (req, res) => {
 
 exports.getAllJuniorAdmins = async (req, res) => {
   try {
+    const token = req.body.token;
+    const userId = jwt.verify(token, process.env.JWT_SECRET);
+    const user = await User.findById(userId.id);
+    if (user.role !== "masterAdmin") {
+      return res.status(400).json({ message: "You are not authorized" });
+    }
     const juniorAdmins = await User.find({ role: "juniorAdmin" });
     res.status(200).json({ juniorAdmins });
   } catch (error) {
